Moved `ImageCommentsPresenter`, view models and localized strings files to production target
