+++ EssentialFeed/EssentialFeed/Image Comments Presentation/ImageCommentsErrorViewModel.swift
@@ -1,0 +1,13 @@
+//
+//  ImageCommentsErrorViewModel.swift
+//  EssentialFeed
+//
+//  Created by Bogdan Poplauschi on 28/03/2021.
+//  Copyright © 2021 Essential Developer. All rights reserved.
+//
+
+import Foundation
+
+public struct ImageCommentsErrorViewModel {
+	public let errorMessage: String?
+}
