+++ EssentialFeed/EssentialFeed/Image Comments Presentation/ImageCommentsLoadingViewModel.swift
@@ -1,0 +1,13 @@
+//
+//  ImageCommentsLoadingViewModel.swift
+//  EssentialFeed
+//
+//  Created by Bogdan Poplauschi on 28/03/2021.
+//  Copyright © 2021 Essential Developer. All rights reserved.
+//
+
+import Foundation
+
+public struct ImageCommentsLoadingViewModel {
+	public let isLoading: Bool
+}
