+++ EssentialFeed/EssentialFeed/Image Comments Presentation/el.lproj/ImageComments.strings
