+++ EssentialFeed/EssentialFeed/Image Comments Presentation/en.lproj/ImageComments.strings
