+++ EssentialFeed/EssentialFeed/Image Comments Presentation/pt-BR.lproj/ImageComments.strings
