+++ EssentialFeed/EssentialFeedTests/Image Comments Presentation/ImageCommentsPresenterTests.swift
@@ -9,105 +9,6 @@
 import EssentialFeed
 import XCTest
 
-struct PresentableImageComment: Equatable {
-	let createdAt: String
-	let message: String
-	let author: String
-}
-
-struct ImageCommentsViewModel {
-	let comments: [PresentableImageComment]
-}
-
-protocol ImageCommentsView {
-	func display(_ viewModel: ImageCommentsViewModel)
-}
-
-struct ImageCommentsLoadingViewModel {
-	let isLoading: Bool
-}
-
-protocol ImageCommentsLoadingView {
-	func display(_ viewModel: ImageCommentsLoadingViewModel)
-}
-
-struct ImageCommentsErrorViewModel {
-	let errorMessage: String?
-}
-
-protocol ImageCommentsErrorView {
-	func display(_ viewModel: ImageCommentsErrorViewModel)
-}
-
-final class ImageCommentsPresenter {
-	private let imageCommentsView: ImageCommentsView
-	private let loadingView: ImageCommentsLoadingView
-	private let errorView: ImageCommentsErrorView
-	private let currentDate: Date
-	
-	public static var title: String {
-		NSLocalizedString(
-			"IMAGE_COMMENTS_VIEW_TITLE",
-			tableName: "ImageComments",
-			bundle: Bundle(for: ImageCommentsPresenter.self),
-			comment: "Title for the image comments view")
-	}
-	
-	private var errorMessage: String {
-		NSLocalizedString(
-			"IMAGE_COMMENTS_VIEW_CONNECTION_ERROR",
-			tableName: "ImageComments",
-			bundle: Bundle(for: ImageCommentsPresenter.self),
-			comment: "Error message when loading comments fails"
-		)
-	}
-	
-	public init(
-		imageCommentsView: ImageCommentsView,
-		loadingView: ImageCommentsLoadingView,
-		errorView: ImageCommentsErrorView,
-		currentDate: Date = Date()
-	) {
-		self.imageCommentsView = imageCommentsView
-		self.loadingView = loadingView
-		self.errorView = errorView
-		self.currentDate = currentDate
-	}
-	
-	public func didStartLoadingComments() {
-		loadingView.display(ImageCommentsLoadingViewModel(isLoading: true))
-		errorView.display(ImageCommentsErrorViewModel(errorMessage: nil))
-	}
-	
-	public func didFinishLoading(with comments: [ImageComment]) {
-		let presentableComments = comments.map { comment in
-			PresentableImageComment(
-				createdAt: relativeDate(from: comment.createdAt),
-				message: comment.message,
-				author: comment.author.username)
-		}
-		imageCommentsView.display(ImageCommentsViewModel(comments: presentableComments))
-		loadingView.display(ImageCommentsLoadingViewModel(isLoading: false))
-	}
-	
-	public func didFinishLoading(with error: Error) {
-		errorView.display(ImageCommentsErrorViewModel(errorMessage: errorMessage))
-		loadingView.display(ImageCommentsLoadingViewModel(isLoading: false))
-	}
-	
-	private static var relativeDateFormatter: RelativeDateTimeFormatter {
-		let formatter = RelativeDateTimeFormatter()
-		formatter.unitsStyle = .full
-		formatter.locale = .current
-		formatter.calendar = Calendar(identifier: .gregorian)
-		return formatter
-	}
-	
-	private func relativeDate(from date: Date) -> String {
-		return ImageCommentsPresenter.relativeDateFormatter.localizedString(for: date, relativeTo: currentDate)
-	}
-}
-
 final class ImageCommentsPresenterTests: XCTestCase {
 	func test_title_isLocalized() {
 		XCTAssertEqual(ImageCommentsPresenter.title, localized("IMAGE_COMMENTS_VIEW_TITLE"))
